moved Object.define to redefine
