--- lib/http-proxy.js
@@ -1,4 +1,0 @@
-(function () {
-
-
-}).call(this);
