+++ lib/http2-XMLHttpRequest.js
@@ -1,0 +1,3 @@
+/**
+ * Created by David on 2/23/17.
+ */
