+++ lib/object-utils.js
@@ -1,0 +1,8 @@
+exports.redefine = function (obj, prop, value) {
+    Object.defineProperty(obj, prop, {
+        enumerable: obj.propertyIsEnumerable(prop),
+        value: value,
+        configurable: true
+    });
+};
+
